+++ Multilingual Excel FileType/Multilingual.Excel.FileType/BatchTasks/ExportBatchTask.cs
@@ -32,7 +32,7 @@ namespace Multilingual.Excel.FileType.BatchTasks
 		GeneratedFileType = AutomaticTaskFileType.BilingualTarget, AllowMultiple = true)]
 	[AutomaticTaskSupportedFileType(AutomaticTaskFileType.BilingualTarget)]
 	[RequiresSettings(typeof(MultilingualExcelExportSettings), typeof(ExportSettingsPage))]
-	public class ExportBatchTask : AbstractFileContentProcessingAutomaticTask
+	public class ExportBatchTask : BatchTaskBase
 	{
 		private MultilingualExcelExportSettings _settings;
 		private ExcelReader _excelReader;
@@ -93,30 +93,6 @@ namespace Multilingual.Excel.FileType.BatchTasks
 				var sdlxliffUpdaterUnProtectEmptyTarget = new SdlxliffUpdater(new EmptyTargetProcessor(_segmentBuilder, false));
 				UpdateFile(projectFile, sdlxliffUpdaterUnProtectEmptyTarget);
 			}
-		}
-
-		private static void UpdateFile(ProjectFile projectFile, SdlxliffUpdater sdlxliffUpdater)
-		{
-			var updatedFilePath = Path.GetTempFileName();
-			sdlxliffUpdater.UpdateFile(projectFile.LocalFilePath, updatedFilePath);
-			File.Copy(updatedFilePath, projectFile.LocalFilePath, true);
-			if (File.Exists(updatedFilePath))
-			{
-				File.Delete(updatedFilePath);
-			}
-		}
-
-		public override bool ShouldProcessFile(ProjectFile projectFile)
-		{
-			var valid = projectFile.FileTypeId == FiletypeConstants.FileTypeDefinitionId;
-			if (!valid)
-			{
-				var message = string.Format(PluginResources.ExceptionMessage_Incorrect_File_Type,
-					projectFile.FileTypeId, FiletypeConstants.FileTypeDefinitionId);
-				throw new Exception(message);
-			}
-
-			return true;
 		}
 
 		public override bool OnFileComplete(ProjectFile projectFile, IMultiFileConverter multiFileConverter)
@@ -238,33 +214,5 @@ namespace Multilingual.Excel.FileType.BatchTasks
 			return projectPath;
 		}
 
-		private List<Models.AnalysisBand> GetAnalysisBands(FileBasedProject project)
-		{
-			var regex = new Regex(@"(?<min>[\d]*)([^\d]*)(?<max>[\d]*)", RegexOptions.IgnoreCase);
-
-			var analysisBands = new List<Models.AnalysisBand>();
-			var type = project.GetType();
-			var internalProjectField = type.GetField("_project", BindingFlags.NonPublic | BindingFlags.Instance);
-			if (internalProjectField != null)
-			{
-				dynamic internalDynamicProject = internalProjectField.GetValue(project);
-				foreach (var analysisBand in internalDynamicProject.AnalysisBands)
-				{
-					Match match = regex.Match(analysisBand.ToString());
-					if (match.Success)
-					{
-						var min = match.Groups["min"].Value;
-						var max = match.Groups["max"].Value;
-						analysisBands.Add(new Models.AnalysisBand
-						{
-							MinimumMatchValue = Convert.ToInt32(min),
-							MaximumMatchValue = Convert.ToInt32(max)
-						});
-					}
-				}
 			}
-
-			return analysisBands;
 		}
-	}
-}
